[BOOK-482] refactor: V2 suffix 제거
