+++ feature/record/src/main/kotlin/com/ninecraft/booket/feature/record/step/EmotionDetailBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.ninecraft.booket.feature.record.step_v2
+package com.ninecraft.booket.feature.record.step
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
